app/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
psycopg2-binary==2.9.9
bcrypt==4.1.3
httpx==0.27.0
python-dotenv==1.0.1
itsdangerous==2.2.0
